SKALE-3726 Update package.json, move tests to unit dir
