+++ test/unit/startUpload.test.js
@@ -6,9 +6,9 @@ require('dotenv').config();
 
 let randomstring = require('randomstring');
 let path = require('path').posix;
-const initFilestorage = require('./utils/helper').initFilestorage;
-const sendTransaction = require('./utils/helper').sendTransaction;
-const getFunds = require('./utils/helper').getFunds;
+const initFilestorage = require('../utils/helper').initFilestorage;
+const sendTransaction = require('../utils/helper').sendTransaction;
+const getFunds = require('../utils/helper').getFunds;
 
 contract('Filestorage', accounts => {
     let filestorage;
