+++ test/unit/uploadChunk.test.js
@@ -5,7 +5,7 @@ chai.should();
 chai.use(require('chai-as-promised'));
 
 let randomstring = require('randomstring');
-const initFilestorage = require('./utils/helper').initFilestorage;
+const initFilestorage = require('../utils/helper').initFilestorage;
 const UPLOADING_GAS = 10 ** 8;
 const CHUNK_LENGTH = 2 ** 10;
